+++ src/lib.rs
@@ -14,9 +14,9 @@ mod signatures;
 pub mod gadgets;
 
 #[deprecated(note = "Please use DoubleSignature instead")]
-pub type Proof = signatures::double_key::Signature;
+pub type Proof = signatures::double::Signature;
 
-pub use keys::public::{PublicKey, PublicKeyPair};
+pub use keys::public::{PublicKey, PublicKeyDouble};
 pub use keys::secret::SecretKey;
-pub use signatures::double_key::Signature as DoubleSignature;
-pub use signatures::single_key::Signature;
+pub use signatures::double::Signature as DoubleSignature;
+pub use signatures::Signature;